backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic[email]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt
openai
youdotcom
tavily-python 